@@ -13,25 +13,25 @@
   },
   "license": "MIT",
   "dependencies": {
-    "@tauri-apps/api": "^2.4.1",
+    "@tauri-apps/api": "^2.5.0",
     "@tauri-apps/plugin-log": "~2.3.1",
     "@tauri-apps/plugin-opener": "^2.2.6",
     "@tauri-apps/plugin-sql": "~2.2.0"
   },
   "devDependencies": {
     "@sveltejs/adapter-static": "^3.0.8",
-    "@sveltejs/kit": "^2.20.4",
+    "@sveltejs/kit": "^2.20.7",
     "@sveltejs/vite-plugin-svelte": "^5.0.3",
-    "@tailwindcss/vite": "^4.1.3",
-    "@tauri-apps/cli": "^2.4.1",
+    "@tailwindcss/vite": "^4.1.4",
+    "@tauri-apps/cli": "^2.5.0",
     "date-fns": "^4.1.0",
     "flowbite": "^3.1.2",
-    "flowbite-svelte": "^0.48.5",
+    "flowbite-svelte": "1.0.0-next.10",
     "flowbite-svelte-icons": "^2.1.1",
-    "svelte": "^5.25.8",
-    "svelte-check": "^4.1.5",
-    "tailwindcss": "^4.1.3",
+    "svelte": "^5.27.0",
+    "svelte-check": "^4.1.6",
+    "tailwindcss": "^4.1.4",
     "typescript": "~5.6.3",
-    "vite": "^6.2.5"
+    "vite": "^6.2.6"
   }
 }
